refactor: slackdeps/http.go -> slackhttp/http.go
