+++ internal/deputil/url.go
@@ -17,11 +17,11 @@ package deputil
 import (
 	"net/http"
 
-	"github.com/slackapi/slack-cli/internal/slackdeps"
+	"github.com/slackapi/slack-cli/internal/slackhttp"
 )
 
 // URLChecker returns url if its status code is 200, otherwise returns empty string
-func URLChecker(httpClient slackdeps.HTTPClient, url string) string {
+func URLChecker(httpClient slackhttp.HTTPClient, url string) string {
 	resp, err := httpClient.Get(url)
 	if err != nil {
 		return ""
