+++ internal/pkg/create/create_test.go
@@ -24,7 +24,7 @@ import (
 	"github.com/slackapi/slack-cli/internal/experiment"
 	"github.com/slackapi/slack-cli/internal/shared"
 	"github.com/slackapi/slack-cli/internal/slackcontext"
-	"github.com/slackapi/slack-cli/internal/slackdeps"
+	"github.com/slackapi/slack-cli/internal/slackhttp"
 	"github.com/spf13/afero"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
@@ -81,14 +81,14 @@ func Test_generateGitZipFileURL(t *testing.T) {
 		templateURL         string
 		gitBranch           string
 		expectedURL         string
-		setupHTTPClientMock func(*slackdeps.HTTPClientMock)
+		setupHTTPClientMock func(*slackhttp.HTTPClientMock)
 	}{
 		"Returns the zip URL using the main branch when no branch is provided": {
 			templateURL: "https://github.com/slack-samples/deno-starter-template",
 			gitBranch:   "",
 			expectedURL: "https://github.com/slack-samples/deno-starter-template/archive/refs/heads/main.zip",
-			setupHTTPClientMock: func(httpClientMock *slackdeps.HTTPClientMock) {
-				res := slackdeps.MockHTTPResponse(http.StatusOK, "OK")
+			setupHTTPClientMock: func(httpClientMock *slackhttp.HTTPClientMock) {
+				res := slackhttp.MockHTTPResponse(http.StatusOK, "OK")
 				httpClientMock.On("Get", mock.Anything).Return(res, nil)
 			},
 		},
@@ -96,8 +96,8 @@ func Test_generateGitZipFileURL(t *testing.T) {
 			templateURL: "https://github.com/slack-samples/deno-starter-template",
 			gitBranch:   "",
 			expectedURL: "https://github.com/slack-samples/deno-starter-template/archive/refs/heads/master.zip",
-			setupHTTPClientMock: func(httpClientMock *slackdeps.HTTPClientMock) {
-				res := slackdeps.MockHTTPResponse(http.StatusOK, "OK")
+			setupHTTPClientMock: func(httpClientMock *slackhttp.HTTPClientMock) {
+				res := slackhttp.MockHTTPResponse(http.StatusOK, "OK")
 				httpClientMock.On("Get", "https://github.com/slack-samples/deno-starter-template/archive/refs/heads/main.zip").Return(nil, fmt.Errorf("HttpClient error"))
 				httpClientMock.On("Get", "https://github.com/slack-samples/deno-starter-template/archive/refs/heads/master.zip").Return(res, nil)
 			},
@@ -106,8 +106,8 @@ func Test_generateGitZipFileURL(t *testing.T) {
 			templateURL: "https://github.com/slack-samples/deno-starter-template",
 			gitBranch:   "pre-release-0316",
 			expectedURL: "https://github.com/slack-samples/deno-starter-template/archive/refs/heads/pre-release-0316.zip",
-			setupHTTPClientMock: func(httpClientMock *slackdeps.HTTPClientMock) {
-				res := slackdeps.MockHTTPResponse(http.StatusOK, "OK")
+			setupHTTPClientMock: func(httpClientMock *slackhttp.HTTPClientMock) {
+				res := slackhttp.MockHTTPResponse(http.StatusOK, "OK")
 				httpClientMock.On("Get", mock.Anything).Return(res, nil)
 			},
 		},
@@ -115,8 +115,8 @@ func Test_generateGitZipFileURL(t *testing.T) {
 			templateURL: "https://github.com/slack-samples/deno-starter-template",
 			gitBranch:   "",
 			expectedURL: "",
-			setupHTTPClientMock: func(httpClientMock *slackdeps.HTTPClientMock) {
-				res := slackdeps.MockHTTPResponse(http.StatusNotFound, "Not Found")
+			setupHTTPClientMock: func(httpClientMock *slackhttp.HTTPClientMock) {
+				res := slackhttp.MockHTTPResponse(http.StatusNotFound, "Not Found")
 				httpClientMock.On("Get", mock.Anything).Return(res, nil)
 			},
 		},
@@ -124,7 +124,7 @@ func Test_generateGitZipFileURL(t *testing.T) {
 			templateURL: "https://github.com/slack-samples/deno-starter-template",
 			gitBranch:   "",
 			expectedURL: "",
-			setupHTTPClientMock: func(httpClientMock *slackdeps.HTTPClientMock) {
+			setupHTTPClientMock: func(httpClientMock *slackhttp.HTTPClientMock) {
 				httpClientMock.On("Get", mock.Anything).Return(nil, fmt.Errorf("HTTPClient error"))
 			},
 		},
@@ -132,7 +132,7 @@ func Test_generateGitZipFileURL(t *testing.T) {
 	for name, tt := range tests {
 		t.Run(name, func(t *testing.T) {
 			// Create mocks
-			httpClientMock := &slackdeps.HTTPClientMock{}
+			httpClientMock := &slackhttp.HTTPClientMock{}
 			tt.setupHTTPClientMock(httpClientMock)
 
 			// Execute
